changed type of score and tetrises arguments in game_ui refresh functions, moved NCursesColors class to new file, added alias to signal type

+++ include/game_ui.h
@@ -9,8 +9,8 @@ class GameUI
 {
 public:
     virtual void refresh_board(const std::vector<std::vector<Pixel>>& pixels) = 0;
-    virtual void refresh_score(int score) = 0;
-    virtual void refresh_tetrises(int tetrises) = 0;
+    virtual void refresh_score(unsigned long long score) = 0;
+    virtual void refresh_tetrises(unsigned long long tetrises) = 0;
     virtual void refresh_next(const Brick& brick) = 0;
     virtual void refresh_hold(const Brick& brick) = 0;
     virtual void game_over() = 0;
